fix: Correct emd_samples bins='auto' with integer inputs on NumPy 2.1+
NumPy 2.1 introduced a constraint that histogram bin widths must be at
least 1 for integer dtypes. This caused emd_samples to create too few
bins for small integer datasets (e.g., emd_samples([1], [2]) returned
0.0 instead of 0.5).

Fix by converting to float64 before computing bin edges, which bypasses
the integer constraint while maintaining identical behavior for float
inputs.

Added regression tests to verify bins='auto' works correctly with
integer inputs.

Fixes #68

+++ src/pyemd/emd.pyx
@@ -212,7 +212,9 @@ def emd_samples(first_array,
         range = (min(np.min(first_array), np.min(second_array)),
                  max(np.max(first_array), np.max(second_array)))
     # Get bin edges using both arrays
-    bins = get_bins(np.concatenate([first_array, second_array]),
+    # Convert to float64 to avoid NumPy 2.1+ integer bin width constraint
+    # (bin width is forced to >= 1 for integer dtypes, causing too few bins)
+    bins = get_bins(np.concatenate([first_array, second_array]).astype(np.float64),
                     range=range,
                     bins=bins)
     # Compute histograms
